Introduce context type to output type type class
The OutputType type class now takes an additional type parameter that is
the context of the schema. Most types work on any context but the object
type needs to have the same context type as the type it is referenced
in. For all other cases a custom type error helps to identify the
problem.