Create support for arguments
Arguments can now be added to fields with the `?>` operator. Arguments
are stored as a record in the field configuration. A type class provides
a function that can turn argument definitions and the list of argument
nodes from the GraphQL query into the parsed argument value. Currently
arguments are always required. We will probably introduce a nullable
wrapper for scalars that makes them nullable. Nullable arguments can
also be left out in the query (instead of providing null).