Create GraphQL.Execution.Result module
This module handles the encoding of results and handling of the errors.
This allows the GraphQL.Type module to simply return results and leave
the JSON encoding to the code handling the results.

Encoding of results and errors is following a defined logic that is
easier to handle and test in the new module.