Create naive implementation for lifted schemas
Schemas can now be lifted into any Monad that supports catching errors
(MonadError). This way a query could be executed synchronously using a
schema parameterised to (Either Error) or asynchronously using Aff. Or
one could implement their own data structure which becomes very relevant
when the n+1 problem needs to be solved. In this case the Applicative
instance will be used for executing neighboring fields and the Bind
instance is used to execute subselections.