Create optionalField and listField functions
These function allow the creation of fields that return lists or fields
that return null. In GraphQL.js this is managed with some extra types
but it becomes really difficult to do that with all the type classes
already around. Therefore I am chosing a design here that is more
similar to how it is done in GraphQL Nexus.