Add logic for catching errors in resolvers
According to the GraphQL specification, errors inside of the resolvers
have to be caught and collected inside of the errors array. The errors
array is then returned in the successful response. We can use
`catchError` on the field execution result to return a ResultError
for the specific field instead of letting the whole execution fail.

See:
- https://graphql.github.io/graphql-spec/June2018/#sec-Errors