Add support for mutations, variables and operation
With this change the `execute` function has been adopted to be able to
execute GraphQL mutations. Furthermore the `execute` function as well as
the `graphql` function now accept variables in a map and an operation
name. This way the query can specify multiple queries and only the one
matching the operation name will be executed.