Change how bind is called before fields
This is a breaking change for the interface of this library. Previously
resolvers were called with the parent's resolver result (meaning with
the result value wrapped in the effect monad itself). This allows for
very efficient passing of the value and the "mapping resolver" is a true
map. Now the resolver is called with the unwrapped value once the value
arrives. I have decided for this interface because it has become
necessary to call >>= on the parent value before we resolve any of the
fields. The reason for this is how Eff and Aff work:

Other than JavaScript's promises, Aff is not executed eagerly. This is
great but also leads to the following:

a = log "a"
b = a >>= const (log "b")
c = a >>= const (log "c")
sequence [b, c]

This code will log: "abac"
The effect is executed twice. We can prevent that by unwrapping a first
in a do block:

do
  a <- log "a"
  let b = const (log "b") a
  let c = const (log "c") a
  sequence [b, c]

Obviously const is not a useful function here but const are the
resolvers in our case. We have to call bind once for all resolvers not
once for every resolver. Otherwise our effects run multiple times. That
is not desireable because we would execute the whole path multiple times
for every leaf node that follows.