Implement union types
With union types we can return different types from a field. Union types
are sum types and on PureScript's side we use variants to dynamically
contruct sum types at compile time. The generated variant type maps to
the record type used to create the union.

Currently inline fragments are ignored by the parser. This has to work
in order to use union types with inline fragments.