Refactor type introspection
The type introspection data type now is a ADT with more specific data
for each possible GraphQL kind. This reduces the amout of boilerplate to
generate these type introspections, but makes returning them from the
resolver in the schema meta object type a bit more verbose. Overall,
this change should still reduce the code and make it more clear, which
type needs to provide which information.