Fix issue in parser with whitespaced queries
The parser was not accepting mutations or queries without the "query"
keyword that were surrounded by whitespace. This was due to the design
of the parser: If a query started with whitespace, it would start
parsing for the "query" keyword and would find a "mutation" keyword. If
we make the whitespace not part of the alternative parsers, it can then
look to parse the "mutation" keyword.