[WIP] Change default resolver
Attemts to change the default resolver to
Record.get (Proxy :: _ name)

This does not work out of the box, because resolvers work a bit like
function composition i.e. the >=> operator. The default resolver is
basically id and then resolvers are stacked on top. But if the default
resolver is something else, that system does not work anymore.

Now the idea would be to change this behaviour. A new resolver would
simply replace the previous resolver. The problem is that a field does
not track the output type. A solution could be to track the output type
of a field.